GitHub actions changes (#21)
* Update publish action to publish to new release as well as maven.

Remove version # defined by gradle.properties. Instead, the publish CI will now automatically generate and tag the commit based on the last version tag. The type of increment is determined by a tag on the PR, one of `major`, `minor`, or `patch`.

If one of these labels is not present CI _will fail_ and no release will be generated!

A few changes have been made outside of source control to this repo for these changes:

1. New labels mentioned above have been added.